+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added support importing from itunes:keywords
 - Added ignoreMainAsset option on fetch from RSS form to prevent fetching large main asset via CURL
 - Now each podcast has two settings pages (general settings, episode settings)
+- Episode summary and description can be different now
 - Improved fetch from RSS via CURL 
 - Improved fetch from RSS progress notifications
 - Improved RSS generation for podcasts